insightbridge4.5 (2)/insightbridge4.5/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database & Persistence
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1

# JWT & Security
pyjwt==2.8.0
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# HTTP Client
httpx==0.25.2

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.12.0
ruff==0.1.8
mypy==1.7.1